Experiment

+++ UserHouse/Controllers/UserController.cs
@@ -3,11 +3,12 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using AutoMapper;
 using UserHouse.Application;
 using UserHouse.Data.Entities;
 using UserHouse.Application.Models;
 using UserHouse.Application.Users;
-using UserHouse.Web.Dtos.Users;
+using UserHouse.Application.Dtos.Users;
 
 namespace UserHouse.Web.Controllers
 {
@@ -15,10 +16,14 @@ namespace UserHouse.Web.Controllers
     [Route("[controller]")]
     public class UserController : Controller
     {
-        private readonly UserAppService _userAppService;
+        private readonly IMapper _mapper;
+        private readonly IUserService _userAppService;
 
-        public UserController(UserAppService userAppService)
+        public UserController(
+            IMapper mapper,
+            IUserService userAppService)
         {
+            _mapper = mapper;
             _userAppService = userAppService;
         }
 
@@ -26,35 +31,67 @@ namespace UserHouse.Web.Controllers
         [Route("Create")]
         public void CreateUser([FromBody] CreateUserDto createUserDto)
         {
-            _userAppService.Create(createUserDto);
+            if (createUserDto != null)
+            {
+                var newUserModel = _mapper.Map<UserModel>(createUserDto);
+
+                _userAppService.Create(newUserModel);
+            }
+            else
+            {
+                throw new Exception("Empty data for creating a new user!");
+            }
         }
 
         [HttpGet]
         [Route("GetById")]
-        public UserModel GetUserById([FromHeader] int userId)
+        public async Task<UserModel> GetUserById([FromHeader] int userId)
         {
-            return _userAppService.GetById(userId);
+            if (userId >= 1)
+            {
+                return await _userAppService.GetById(userId);
+            }
+            else
+            {
+                throw new Exception("Given id is not valid!");
+            }
         }
 
         [HttpGet]
         [Route("GetAll")]
-        public List<UserModel> GetAllUsers()
+        public async Task<List<UserModel>> GetAllUsers()
         {
-            return _userAppService.GetAll();
+            return await _userAppService.GetAll();
         }
 
         [HttpPut]
         [Route("Update")]
         public void UpdateUser([FromBody] UserDto userDto)
         {
-            _userAppService.Update(userDto);
+            if (userDto != null)
+            {
+                var user = _mapper.Map<UserModel>(userDto);
+
+                _userAppService.Update(user);
+            }
+            else
+            {
+                throw new Exception("Empty data for updating a user!");
+            }
         }
 
         [HttpDelete]
         [Route("Delete")]
         public void DeleteUser([FromHeader] int userId)
         {
-            _userAppService.Delete(userId);
+            if (userId >= 1)
+            {
+                _userAppService.Delete(userId);
+            }
+            else
+            {
+                throw new Exception("Given id is not valid!");
+            }
         }
     }
 }
